api/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.1
httpx>=0.26.0

# Database
chromadb>=0.6.0

# AI Integration
openai>=1.10.0
anthropic>=0.21.3

# Document Processing
langchain>=0.2.0
langchain-community>=0.2.0
tiktoken>=0.7.0

# Markdown Processing
markdown>=3.5.0
beautifulsoup4>=4.12.0
tavily-python>=0.3.0
